CDPS-1858: Refactor 14/14 - globalEmailsController.ts
